+++ src/java/Operator/OperatorVO.java
@@ -16,7 +16,7 @@ public class OperatorVO {
     private int programId;
     private String operatorDate;
     private String operatorDetailDes;
-    private double actualExpense;
+    private long actualExpense;
     private List<OperatorImage> activiesImgs;
     private List<OperatorImage> billImgs;
     private String city;
